add `Ctrl+Click` and `Ctrl+A` select, improve tabindex for collections

+++ client/src/components/History/Content/ContentItem.vue
@@ -66,7 +66,7 @@
                     :is-visible="item.visible"
                     :state="state"
                     :item-urls="itemUrls"
-                    :keyboard-selectable="expandDataset"
+                    :keyboard-selectable="isCollection || expandDataset"
                     @delete="onDelete"
                     @display="onDisplay"
                     @showCollectionInfo="onShowCollectionInfo"
@@ -243,19 +243,23 @@ export default {
             } else if (event.key === "ArrowUp" || event.key === "ArrowDown") {
                 event.preventDefault();
                 this.$emit("arrow-navigate", event.key);
-            } else if (event.key === "Delete") {
+            } else if (event.key === "Delete" && !this.selected && !this.item.deleted) {
                 event.preventDefault();
                 this.onDelete(event.shiftKey);
             } else if (event.key === "Escape") {
                 event.preventDefault();
                 this.$emit("hide-selection");
+            } else if (event.key === "a" && event.ctrlKey) {
+                event.preventDefault();
+                this.$emit("select-all");
             }
         },
-        onClick() {
-            if (this.isPlaceholder) {
+        onClick(event) {
+            if (event && event.ctrlKey) {
+                this.$emit("update:selected", !this.selected);
+            } else if (this.isPlaceholder) {
                 return;
-            }
-            if (this.isDataset) {
+            } else if (this.isDataset) {
                 this.$emit("update:expand-dataset", !this.expandDataset);
             } else {
                 this.$emit("view-collection", this.item, this.name);
